util: specify integer return type for function "get_num_logical_cores"

+++ src/common/util.c
@@ -37,7 +37,7 @@
 
 #include "util.h"
 
-int get_num_logical_cores(void)
+uint16_t get_num_logical_cores(void)
 {
 #ifdef _WIN32
   SYSTEM_INFO sysinfo;
